iot_for_decisionmaking/requirements.txt
influxdbr
shiny
shinydashboard
shinyjs
ggplot2
DT
data.table
tinytex
plotly
openssl
knitr
rmarkdown
